refactor:create utils package and add  AppDestination to it
